requirements.txt
torch
torchvision
scikit-learn
pandas
numpy
xgboost
streamlit
pillow
joblib